separated `data.py` for core part and file I/O part.
rename:
  data -> dataset
  DataGenerator -> AtomicStructureDatasetGenerator
